--- Bisaya++.html
@@ -1,7 +1,0 @@
-<!DOCTYPE html>
-<html>
-    <head>
-        <title>Bisaya++</title>
-        <script src = "Bisaya++.js"></script>
-    </head>
-</html>
