--- README.md
@@ -1,26 +1,0 @@
-TO DO:
-modulo operator
-If else
-for loop
-escape sequence - []
-
-
-fix later:
-rearding the & and $. when placed in a character or string should be treated as a concatination or nextline
-comments should be put anywhere
-
-Note - Go to Bisaya++copy.js
-
-DONE:
-variable declaration
-variable initialization
-ipakita
-comments
-DAWATA
-arithmetic
-Logical expressions and comparison operator
-
-
-
-Packages need to download:
-readline-sync - npm install readline-sync
