+++ arithmetic/NumberNode.js
@@ -1,0 +1,12 @@
+class NumberNode {
+    constructor(tok) {
+        this.tok = tok;
+        this.pos_start = this.tok.pos_start
+        this.pos_end = this.tok.pos_end
+    }
+
+    toString() {
+        return `${this.tok}`;
+    }
+}
+module.exports = { NumberNode };
